requirements.txt
mcp>=1.2.0
langchain==0.2.7
langgraph>=0.2.24
langchain-openai>=0.2.3
langchain-mcp-adapters>=0.0.7
pydantic>=2.7
streamlit>=1.36
python-dotenv>=1.0
requests>=2.32.3